refactor: 모델 파일 구조 개선 - AI 모델들을 alimtalk_models.py로 통합
- ai_models.py 파일 제거하여 모델 관리 단순화
- ChatRequest, ChatResponse, TemplateModificationRequest, TemplateModificationResponse를 alimtalk_models.py로 이동
- ai_routes.py에서 import 경로를 alimtalk_models로 변경
- 모든 모델을 한 곳에서 관리하여 프로젝트 구조 개선
